SafeBox File Manger/requirements.txt
pycryptodome>=3.10
